Campo price de product -> Decimal -> Int
Campo price alterado para int para uma maior precisão em calculos extras como imposto, etc.

+++ migration/src/m20250903_200616_create_products_table.rs
@@ -15,7 +15,7 @@ impl MigrationTrait for Migration {
                     .col(pk_auto(Product::Id))
                     .col(ColumnDef::new(Product::Name).string().not_null())
                     .col(ColumnDef::new(Product::Description).string().null())
-                    .col(ColumnDef::new(Product::Price).decimal().not_null())
+                    .col(ColumnDef::new(Product::Price).integer().not_null())
                     .col(
                         ColumnDef::new(Product::Sku)
                             .string()
